Added tower icons and improvements
Added tower icons for the tower buttons and appropriate behaviors for being clicked.
Added icon for pause menu.
Added button for returning to menu.
Added quit button to menu.
Added proper ending to tutorial.
Adjusted a bit of code in the resource_generator so that it works in multiple scenes instead of being hard-coded to basic level.